Separate helper functions to not need to export test obj in files
