+++ lib/configure.js
@@ -1,14 +1,14 @@
 'use strict'
 
-const { openSync, closeSync, promises: fs } = require('graceful-fs')
+const { promises: fs } = require('graceful-fs')
 const path = require('path')
 const log = require('./log')
 const os = require('os')
 const processRelease = require('./process-release')
 const win = process.platform === 'win32'
 const findNodeDirectory = require('./find-node-directory')
-const createConfigGypi = require('./create-config-gypi')
-const { format: msgFormat } = require('util')
+const { createConfigGypi } = require('./create-config-gypi')
+const { format: msgFormat, findAccessibleSync } = require('util')
 const { findPython } = require('./find-python')
 const { findVisualStudio } = win ? require('./find-visualstudio') : {}
 
@@ -277,32 +277,5 @@ async function configure (gyp, argv) {
   }
 }
 
-/**
- * Returns the first file or directory from an array of candidates that is
- * readable by the current user, or undefined if none of the candidates are
- * readable.
- */
-function findAccessibleSync (logprefix, dir, candidates) {
-  for (let next = 0; next < candidates.length; next++) {
-    const candidate = path.resolve(dir, candidates[next])
-    let fd
-    try {
-      fd = openSync(candidate, 'r')
-    } catch (e) {
-      // this candidate was not found or not readable, do nothing
-      log.silly(logprefix, 'Could not open %s: %s', candidate, e.message)
-      continue
-    }
-    closeSync(fd)
-    log.silly(logprefix, 'Found readable %s', candidate)
-    return candidate
-  }
-
-  return undefined
-}
-
 module.exports = configure
-module.exports.test = {
-  findAccessibleSync
-}
 module.exports.usage = 'Generates ' + (win ? 'MSVC project files' : 'a Makefile') + ' for the current module'
