+++ lib/create-config-gypi.js
@@ -143,8 +143,8 @@ async function createConfigGypi ({ gyp, buildDir, nodeDir, vsInfo, python }) {
   return configPath
 }
 
-module.exports = createConfigGypi
-module.exports.test = {
+module.exports = {
+  createConfigGypi,
   parseConfigGypi,
   getCurrentConfigGypi
 }
