+++ lib/util.js
@@ -1,8 +1,9 @@
 'use strict'
 
-const log = require('./log')
 const cp = require('child_process')
 const path = require('path')
+const { openSync, closeSync } = require('graceful-fs')
+const log = require('./log')
 
 const execFile = async (...args) => new Promise((resolve) => {
   const child = cp.execFile(...args, (...a) => resolve(a))
@@ -48,8 +49,33 @@ async function regSearchKeys (keys, value, addOpts) {
   }
 }
 
+/**
+ * Returns the first file or directory from an array of candidates that is
+ * readable by the current user, or undefined if none of the candidates are
+ * readable.
+ */
+function findAccessibleSync (logprefix, dir, candidates) {
+  for (let next = 0; next < candidates.length; next++) {
+    const candidate = path.resolve(dir, candidates[next])
+    let fd
+    try {
+      fd = openSync(candidate, 'r')
+    } catch (e) {
+      // this candidate was not found or not readable, do nothing
+      log.silly(logprefix, 'Could not open %s: %s', candidate, e.message)
+      continue
+    }
+    closeSync(fd)
+    log.silly(logprefix, 'Found readable %s', candidate)
+    return candidate
+  }
+
+  return undefined
+}
+
 module.exports = {
   execFile,
   regGetValue,
-  regSearchKeys
+  regSearchKeys,
+  findAccessibleSync
 }
